# Frontend/openapi/user.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "이메일 인증",
      "description": "LOCAL 회원가입용 이메일 인증 코드 요청·확인"
    }
  ],
  "paths": {
    "/api/v1/auth/email-verification": {
      "post": {
        "tags": ["이메일 인증"],
        "summary": "이메일 인증 코드 요청",
        "description": "가입 여부와 관계없이 같은 응답이며, 메일 발송 완료를 기다리지 않는다. 발송 제한은 MEMBER_AUTH.md 1.2 참고",
        "operationId": "requestCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "요청 처리 완료"
          },
          "400": {
            "description": "INVALID_REQUEST, VALIDATION_FAILED, INVALID_EMAIL"
          },
          "429": {
            "description": "EMAIL_VERIFICATION_RESEND_TOO_SOON"
          }
        }
      }
    },
    "/api/v1/auth/email-verification/confirm": {
      "post": {
        "tags": ["이메일 인증"],
        "summary": "이메일 인증 코드 확인",
        "description": "성공하면 15분간 유효한 가입 토큰을 email_signup_token HttpOnly 쿠키로 발급한다. 코드 형식 오류는 시도 횟수에 포함하지 않는다",
        "operationId": "confirmCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "확인 성공, Set-Cookie로 가입 토큰 발급"
          },
          "400": {
            "description": "INVALID_REQUEST, VALIDATION_FAILED, INVALID_EMAIL, EMAIL_VERIFICATION_CODE_MISMATCH, EMAIL_VERIFICATION_CODE_EXPIRED"
          },
          "409": {
            "description": "DUPLICATE_EMAIL"
          },
          "429": {
            "description": "EMAIL_VERIFICATION_ATTEMPTS_EXCEEDED"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmailVerificationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["email"]
      },
      "EmailVerificationConfirmRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string",
            "pattern": "[0-9]{6}"
          }
        },
        "required": ["code", "email"]
      }
    }
  }
}